Removed circular dependency.
